https://github.com/mulesoft/api-workbench/issues/247 - initial implementation, https://github.com/mulesoft/api-workbench/issues/268 fixed.
