+++ src/atom/context-menu/commonContextActions.ts
@@ -3,199 +3,35 @@
 import _ = require("underscore")
 import contextMenu = require("./contextMenu")
 import commandManager = require("../quick-commands/command-manager")
-import contextActions = require("./contextActions")
+import actions = require("raml-actions")
 import path = require ('path')
 import provider=require("../suggestion/provider")
 import rp=require("raml-1-parser");
 import search=rp.search;
 import hl=rp.hl;
+import ll=rp.ll;
 import editorTools = require("../editor-tools/editor-tools")
 
-function registerCommonActions() {
-    //TODO here the common actions should be registered
-}
-
-/**
- * Must be called once on module initialization
- */
-export function initialize() {
-
-    initializeActionSupport();
-
-    registerCommonActions()
-}
-
-/**
- * General state of AST of the opened editor
- */
-export interface IGeneralASTState {
-    editor : AtomCore.IEditor
-    offset : number
-    node : hl.IParseResult
-    completionKind : search.LocationKind
-}
-
-/**
- * For those who ignore state calculators approach. Is not recommended to use.
- */
-export class NullCalculator implements contextActions.IContextStateCalculator {
-
-    calculate () : any {
-    }
-}
-
-/**
- * This class calculates current open editor AST state, including the selected node.
- *
- * The state is actually calculated on the global calculation start, and calling "calculate"
- * just returns the state. This allows to reuse a single instance in many actions
- * and only perform the actual state calculation once.
- *
- * On reuse please call contextCalculationStarted and contextCalculationFinished methods
- * from respective methods of the state calculator that reuses current one.
- *
- * It is not recommended to inherit the class, instead, reuse the exported instance of the class
- * so that AST parsing is performed once.
- */
-export class GeneralASTStateCalculator implements contextActions.IContextStateCalculator {
-
-    private state : IGeneralASTState = null;
-
-    /**
-     * Is called to calculate context
-     */
-    calculate () : any {
-
-        //should actually never happened if this class is reused properly
-        if (this.state == null) {
-            this.state = this.calculateState()
-        }
-
-        return this.state
-    }
-
-    /**
-     * If present is called before any context calculations are started
-     */
-    contextCalculationStarted : () => void = () => {
-
-        if (this.state == null) {
-            this.state = this.calculateState()
-        }
-    }
-
-    /**
-     * If present is called after all context calculations are finished
-     */
-    contextCalculationFinished : () => void = () => {
-
-        //deleting current state
-        this.state = null
-    }
-
-    private calculateState() : IGeneralASTState {
-
-        var gotEditorFromOutline = false;
-
-        var editor = null;
-        if (atom.workspace.getActiveTextEditor()) {
-            editor = atom.workspace.getActiveTextEditor()
-        } else if (editorTools.aquireManager()) {
-            editor = <AtomCore.IEditor>editorTools.aquireManager().getCurrentEditor()
-            gotEditorFromOutline = true
-        }
-
-        if (!editor) return null
-
-        if (path.extname(editor.getPath()) != '.raml') return null
-
-        var request = {
-            editor: editor,
-            bufferPosition: editor.getCursorBufferPosition()
-        };
-
-        var node = null;
-
-        if (gotEditorFromOutline) {
-            node = editorTools.aquireManager().getSelectedNode()
-        } else {
-            if (editor.getBuffer()) {
-                var lastPosition = editor.getBuffer().getEndPosition();
-                if (lastPosition.column == request.bufferPosition.column
-                    && lastPosition.row == request.bufferPosition.row) {
-                    return null;
-                }
-                if (request.bufferPosition.row == 0 && request.bufferPosition.column == 0) {
-                    return null;
-                }
-            }
-
-            node = provider.getAstNode(request, false);
-        }
-
-        if (!node) {
-            return null;
-        }
-
-        var offset = request.editor.getBuffer().
-            characterIndexForPosition(request.bufferPosition);
-
-        var completionKind = search.determineCompletionKind(editor.getBuffer().getText(), offset);
-
-        return {
-            editor : editor,
-            offset : offset,
-            node : node,
-            completionKind : completionKind
-        }
-    }
-}
-
-export var generalASTStateCalculator = new GeneralASTStateCalculator()
-
-/**
- * Intended for subclassing version of GeneralASTStateCalculator
- * Override calculate() method, use getGeneralState() to obtain current general AST state
- */
-export class CommonASTStateCalculator  implements contextActions.IContextStateCalculator {
-
-    calculate () : any {
-       return null
-    }
-
-    getGeneralState() : IGeneralASTState {
-        return <IGeneralASTState> generalASTStateCalculator.calculate()
-    }
-
-    contextCalculationStarted : () => void = () => {
-        generalASTStateCalculator.contextCalculationStarted()
-    }
-
-    contextCalculationFinished : () => void = () => {
-        generalASTStateCalculator.contextCalculationFinished()
-    }
-
-    getEditor() : AtomCore.IEditor {
-        return editorTools.aquireManager()?(<AtomCore.IEditor>editorTools.aquireManager().getCurrentEditor()) : atom.workspace.getActiveTextEditor();
-    }
-}
+export type CommonASTStateCalculator = actions.CommonASTStateCalculator;
 
 function initializeActionSupport() {
 
-    var editorContextMenuContributor : contextMenu.IContextMenuContributor = {
+    actions.intializeStandardActions();
+
+    var editorContextMenuContributor : actions.IContextMenuContributor = {
 
         id : "editorContextActionContributor",
 
 
         calculateItems : function () {
-            var actions = contextActions.calculateCurrentActions(
-                contextActions.TARGET_RAML_EDITOR_NODE)
+            var currentActions = actions.calculateCurrentActions(
+                actions.TARGET_RAML_EDITOR_NODE)
 
-            if (!actions) return []
+            if (!currentActions) return []
 
-            var result : contextMenu.IContextMenuItem[] = []
+            var result : actions.IContextMenuItem[] = []
 
-            actions.forEach(action => {
+            currentActions.forEach(action => {
                 result.push({
 
                     selector : 'atom-text-editor[data-grammar="source raml"],.raml-outline',
@@ -204,7 +40,9 @@ function initializeActionSupport() {
 
                     categories : action.category,
 
-                    onClick: action.onClick
+                    onClick: action.onClick,
+
+                    children: []
                 })
             })
 
@@ -213,15 +51,15 @@ function initializeActionSupport() {
 
     }
 
-    contextMenu.registerContributor(editorContextMenuContributor)
+    actions.registerContributor(editorContextMenuContributor)
 
     var editorCommandContributor : commandManager.ICommandContributor = {
         id : "editorContextActionContributor",
 
 
         calculateItems : function () {
-            var actions = contextActions.calculateCurrentActions(
-                contextActions.TARGET_RAML_EDITOR_NODE)
+            var actions = actions.calculateCurrentActions(
+                actions.TARGET_RAML_EDITOR_NODE)
 
             if (!actions) return []
 
@@ -243,4 +81,103 @@ function initializeActionSupport() {
     }
 
     commandManager.registerContributor(editorCommandContributor)
+}
+
+export function initialize() {
+    var editorProvider = {
+        getCurrentEditor() {
+            var gotEditorFromOutline = false;
+
+            var editor = null;
+
+            if (atom.workspace.getActiveTextEditor()) {
+                editor = atom.workspace.getActiveTextEditor()
+            } else if (editorTools.aquireManager()) {
+                editor = <AtomCore.IEditor>editorTools.aquireManager().getCurrentEditor()
+                gotEditorFromOutline = true
+            }
+
+            return editor;
+        }
+    }
+
+    actions.setEditorProvider(editorProvider);
+
+    var astProvider = {
+            getASTRoot() : hl.IHighLevelNode {
+                return this.getSelectedNode().root();
+            },
+
+            getSelectedNode() : hl.IParseResult {
+
+                var editor = null;
+                var gotEditorFromOutline = false;
+
+                if (editorTools.aquireManager()) {
+                    editor = <AtomCore.IEditor>editorTools.aquireManager().getCurrentEditor()
+                    gotEditorFromOutline = true
+                }
+
+                if (!editor && atom.workspace.getActiveTextEditor()) {
+                    editor = atom.workspace.getActiveTextEditor()
+                }
+
+                if (!editor) return null
+
+                if (path.extname(editor.getPath()) != '.raml') return null
+
+                var request = {
+                    editor: editor,
+                    bufferPosition: editor.getCursorBufferPosition()
+                };
+
+                var node = null;
+
+                if (gotEditorFromOutline) {
+                    node = editorTools.aquireManager().getSelectedNode()
+                } else {
+                    if (editor.getBuffer()) {
+                        var lastPosition = editor.getBuffer().getEndPosition();
+                        if (lastPosition.column == request.bufferPosition.column
+                            && lastPosition.row == request.bufferPosition.row) {
+                            return null;
+                        }
+                        if (request.bufferPosition.row == 0 && request.bufferPosition.column == 0) {
+                            return null;
+                        }
+                    }
+
+                    node = provider.getAstNode(request, false);
+                }
+
+                return node;
+            }
+    }
+
+    actions.setASTProvider(<any>astProvider);
+
+    var astModifier = {
+        deleteNode(node: hl.IParseResult) {
+            var editorManager = editorTools.aquireManager();
+            if (editorManager && editorManager._view) {
+                editorManager._view.forEachViewer(x=> x.remove(node));
+            }
+
+            var parent = node.parent();
+            if (parent) {
+                parent.remove(<any>node);
+                parent.resetChildren();
+            }
+        },
+        updateText(node: ll.ILowLevelASTNode) {
+            var editorManager = editorTools.aquireManager();
+            if (editorManager) {
+                editorManager.updateText(node);
+            }
+        }
+    }
+
+    actions.setASTModifier(<any>astModifier);
+
+    initializeActionSupport();
 }
