+++ src/atom/context-menu/quickFix.ts
@@ -2,9 +2,12 @@
 
 import fs = require ('fs')
 import path = require ('path')
-import rp = require("raml-1-parser")
+import contextActions = require("raml-actions")
+import parser2 = require("raml-1-parser")
+import rp = contextActions.parser
 import lowLevel=rp.ll;
 import hl=rp.hl;
+import hl2=parser2.hl;
 import search=rp.search;
 import stubs=rp.stubs;
 import universeHelpers = rp.universeHelpers;
@@ -20,12 +23,19 @@ import SpacePenViews = require('atom-space-pen-views')
 import def=rp.ds
 import move=require("../dialogs/moveElementsDialog")
 import tooltip=require("../core/tooltip-manager")
-import contextActions = require("./contextActions")
 import commonContextActions = require("./commonContextActions")
 import assistUtils = require("../dialogs/assist-utils")
 import textutil = require("../../util/textutil")
 import editorTools=require("../editor-tools/editor-tools")
 
+/**
+ * For unknown reason, compiler cant merge highlevel coming from raml-1-parser direct import
+ * and the one coming from actions, so we need to convert manually.
+ * @param node
+ */
+function hlConv(node:hl.IHighLevelNode) : hl2.IHighLevelNode {
+    return <any> node;
+}
 
 export class AbstractDialogWithValidation {
 
@@ -205,7 +215,7 @@ export class AbstractDialogWithValidation {
 }
 
 
-class MoveResourceStateCalculator extends commonContextActions.CommonASTStateCalculator {
+class MoveResourceStateCalculator extends contextActions.CommonASTStateCalculator {
 
     calculate():any {
 
@@ -225,7 +235,7 @@ class MoveResourceStateCalculator extends commonContextActions.CommonASTStateCal
     }
 }
 
-class CreateGlobalSchemaStateCalculator extends commonContextActions.CommonASTStateCalculator {
+class CreateGlobalSchemaStateCalculator extends contextActions.CommonASTStateCalculator {
 
     calculate():any {
 
@@ -433,7 +443,7 @@ export function createGlobalSchema(attr:hl.IAttribute) {
     new CreateGlobalSchemaDialog(attr, "Create Global Schema").show()
 }
 
-class ExpandSignatureStateCalculator extends commonContextActions.CommonASTStateCalculator {
+class ExpandSignatureStateCalculator extends contextActions.CommonASTStateCalculator {
 
     calculate():any {
 
@@ -467,327 +477,12 @@ function updateEditor(node) {
     ed.setText(assistUtils.cleanEmptyLines(node.root().lowLevel().unit().contents()));
 }
 
-var getKeyValue = function (offset, txt) {
-    var m = offset;
-
-    for (var i = offset; i >= 0; i--) {
-        var c = txt.charAt(i);
-        if (c == ' ' || c == '\r' || c == '\n' || c == '\t') {
-            m = i + 1;
-            break;
-        }
-    }
-    var res = "";
-    for (var i = m; i < txt.length; i++) {
-        var c = txt.charAt(i);
-        if (c == ' ' || c == '\r' || c == '\n' || c == '\t' || c == ':') {
-            break;
-        }
-        res += c;
-    }
-    return res;
-};
-
-class CompleteBodyStateCalculator extends commonContextActions.CommonASTStateCalculator {
-
-    calculate():any {
-
-        var generalState = this.getGeneralState()
-        if (!generalState) return null
-
-        if (generalState.completionKind != search.LocationKind.KEY_COMPLETION)
-            return null;
-
-        var highLevelNode = <hl.IHighLevelNode>generalState.node;
-
-        if (universeHelpers.isResponseType(highLevelNode.definition()) ||
-            universeHelpers.isMethodType(highLevelNode.definition())) {
-            var txt = generalState.editor.getText();
-            var res = getKeyValue(generalState.offset, txt);
-            if (res == "body") {
-                return highLevelNode
-            }
-        }
-        if (universeHelpers.isBodyLikeType(highLevelNode.definition())) {
-            if (highLevelNode.elements().length == 0) {
-                return highLevelNode
-            }
-        }
-
-        return null
-    }
-}
-
-export function saveExample(r:hl.IHighLevelNode, schp:string, content:string) {
-    var ed = assistUtils.getActiveEditor();
-    var sdir = path.resolve(path.dirname(ed.getPath()), path.dirname(schp));
-    if (!fs.existsSync(sdir)) {
-        fs.mkdirSync(sdir);
-    }
-    var shFile = path.resolve(path.dirname(ed.getPath()), schp);
-    fs.writeFileSync(shFile, content)
-}
-
-class FillBodyDialog {
-
-    protected name:string = ""
-
-    constructor(private h:hl.IHighLevelNode, private body:hl.IHighLevelNode, protected title:string = "Fill body") {
-
-    }
-
-    extraContent(s:UI.Section) {
-
-    }
-
-    needXML:boolean = true;
-    needJSON:boolean = true;
-    createButton:UI.Button;
-
-    updateButtons() {
-        if (!this.createButton) {
-            return;
-        }
-        if (this.name.length == 0) {
-            this.createButton.setDisabled(true);
-            this.em.setDisplay(true)
-            this.em.setText("Please type name of your payload");
-            return;
-        }
-        if (this.needJSON) {
-            try {
-                JSON.parse(this.jsexample);
-            } catch (e) {
-                this.createButton.setDisabled(true);
-                this.em.setDisplay(true)
-                this.em.setText("JSON example is not correct");
-                return;
-            }
-            try {
-                var so = su.getJSONSchema(this.jsschema, null);
-
-            } catch (e) {
-                this.createButton.setDisabled(true);
-                this.em.setDisplay(true)
-                this.em.setText("JSON schema is not correct");
-                return;
-            }
-        }
-        if (this.needXML) {
-            try {
-                xmlutil(this.xmlexample);
-            } catch (e) {
-                this.createButton.setDisabled(true);
-                this.em.setDisplay(true)
-                this.em.setText("XML example is not correct");
-                return;
-            }
-            try {
-                var so = su.getXMLSchema(this.xmlschema);
-
-            } catch (e) {
-                this.createButton.setDisabled(true);
-                this.em.setDisplay(true)
-                this.em.setText("XML schema is not correct");
-                return;
-            }
-        }
-        this.em.setDisplay(false);
-        this.createButton.setDisabled(false);
-    }
-
-    em:UI.Label;
-
-    show() {
-        var zz = null;
-        this.em = UI.label("Please type name of your payload", UI.Icon.BUG, UI.TextClasses.ERROR, UI.HighLightClasses.NONE);
-        var section = UI.section(this.title, UI.Icon.BOOK, false, false, this.em, UI.h3("Please type name for your payload")).pad(10, 10)
-        section.addChild(UI.texfField("", this.name, x=> {
-            this.name = x.getBinding().get();
-            this.updateButtons();
-        }))
-        var r1 = UI.checkBox("Create XML body");
-        r1.setValue(this.needXML);
-        r1.getBinding().addListener(x=> {
-            this.needXML = r1.getValue();
-            this.updateButtons();
-        });
-        section.addChild(r1);
-        var r2 = UI.checkBox("Create JSON body");
-        r2.setValue(this.needJSON);
-        r2.getBinding().addListener(x=> {
-            this.needJSON = r2.getValue();
-            this.updateButtons();
-        });
-        section.addChild(r2);
-
-        var buttonBar = UI.hc().setPercentWidth(100).setStyle("display", "flex");
-        buttonBar.addChild(UI.label("", null, null, null).setStyle("flex", "1"))
-        buttonBar.addChild(UI.button("Cancel", UI.ButtonSizes.NORMAL, UI.ButtonHighlights.NO_HIGHLIGHT, UI.Icon.NONE, x=> {
-            zz.destroy()
-        }).margin(10, 10))
-
-        this.createButton = UI.button("Create", UI.ButtonSizes.NORMAL, UI.ButtonHighlights.SUCCESS, UI.Icon.NONE, x=> {
-            this.onOk(zz);
-            zz.destroy();
-        });
-        buttonBar.addChild(this.createButton)
-        var tf = new UI.TabFolder();
-        this.createButton.setDisabled(true)
-        this.createTextSection(tf, "JSON Example", "source.json", "jsexample");
-        this.createTextSection(tf, "JSON Schema", "source.json", "jsschema");
-        this.createTextSection(tf, "XML Example", "text.xml", "xmlexample");
-        this.createTextSection(tf, "XML Schema", "text.xml", "xmlschema");
-        tf.setOnSelected(()=> {
-            var c = tf.selectedComponent();
-            var te = (<UI.AtomEditorElement><any>c.children()[1]);
-            te.setText((<any>this)[(<UI.BasicComponent<any>>c).id()]);
-
-        })
-        section.addChild(tf);
-        section.addChild(buttonBar);
-        zz = (<any>atom).workspace.addModalPanel({item: section.renderUI()});
-    }
-
-    jsexample:string = '{\n "message":"Hello world"\n}'
-    xmlexample:string = "";
-    xmlschema:string = "";
-    jsschema:string = "";
-
-    private createTextSection(tf:UI.TabFolder, caption:string, lang:string, code:string) {
-        var hs = UI.vc();
-        hs.setCaption(caption)
-        hs.setId(code)
-        var ts = new UI.AtomEditorElement("", x=>x);
-        ts.setMini(false);
-        ts.getBinding().addListener(x=> {
-            this[code] = ts.getValue();
-            this.updateButtons()
-        })
-        //ts.setCaption(code)
-        ts.setText("" + (<any>this)[code]);
-        ts.setCaption(caption)
-        ts.setGrammar(lang)
-        ts.setStyle("height", "400px");
-        ts.setStyle("border", "solid");
-        ts.setStyle("border-width", "1px");
-        hs.addChild(UI.h3("Please type your example here:"))
-
-        hs.addChild(ts);
-        if (code == 'jsexample') {
-            var b = UI.button("Generate JSON schema", UI.ButtonSizes.NORMAL, UI.ButtonHighlights.SUCCESS, UI.Icon.NONE, x=> {
-                try {
-                    var rs = shemagen.generateSchema(this.jsexample, "application/json")
-                    this.jsschema = rs;
-                    tf.setSelectedIndex(1)
-                }
-                catch (e) {
-                    this.jsschema = e.message;
-                    tf.setSelectedIndex(1)
-                }
-            });
-            hs.addChild(b.margin(5, 5, 5, 5));
-        }
-        if (code == 'xmlexample') {
-            var b = UI.button("Generate JSON example", UI.ButtonSizes.NORMAL, UI.ButtonHighlights.SUCCESS, UI.Icon.NONE, x=> {
-                try {
-                    var rs = xmlutil(this.xmlexample)
-                    this.jsexample = JSON.stringify(rs, null, 2);
-                    tf.setSelectedIndex(0)
-                }
-                catch (e) {
-                    this.jsexample = e.message;
-                    tf.setSelectedIndex(0)
-                }
-            });
-            hs.addChild(b.margin(5, 5, 5, 5));
-        }
-        tf.add(caption, null, hs);
-    }
 
 
-    protected onOk(zz) {
-        var bodyType = <def.NodeClass>this.body.definition().universe().type("BodyLike");
-        if (bodyType) {
-            //RAML 0.8 case
 
-            var node = this.body;
-            if (universeHelpers.isBodyProperty(node.property())) {
-                node = node.parent();
-            }
-            var type = <def.NodeClass>node.definition();
 
-            if (this.needJSON) {
-                var body = <hl.IEditableHighLevelNode>stubs.createStubNode(bodyType,type.property('name'), "application/json");
-                body.createAttr("schema", this.name);
-                body.createAttr("example", "!include ./examples/" + this.name + ".json");
-                node.add(body);
-                assistUtils.createGlobalSchemaFromNameAndContent(this.h.root(), this.name, "schemas/" + this.name + ".json", this.jsschema)
-                saveExample(this.h, "./examples/" + this.name + ".json", this.jsexample);
-            }
-            if (this.needXML) {
-                var body = <hl.IEditableHighLevelNode>stubs.createStubNode(bodyType,type.property('name'), "application/xml");
-                body.createAttr("schema", this.name + "-xml");
-                body.createAttr("example", "!include ./examples/" + this.name + ".xml");
-                node.add(body);
-                var xmlSchemaContents = this.xmlschema;
-                assistUtils.createGlobalSchemaFromNameAndContent(this.h.root(), this.name + "-xml", "schemas/" + this.name + ".xml", xmlSchemaContents)
-                saveExample(this.h, "./examples/" + this.name + ".xml", this.xmlexample);
-            }
-        } else {
-            //RAML 1.0 case
-            var response = this.body;
-            if (!universeHelpers.isResponseType(response.property().range())) {
-                console.log("Incorrect parent " + response.printDetails() + " , expecting response")
-                return;
-            }
 
-            var responseWrapper = <wrapper.Response>response.wrapperNode();
-
-            var bodies: wrapper.TypeDeclaration[] = [];
-
-            if (this.needJSON) {
-                var typeName = "application/json"
-
-                var bodyWrapper = apiModifier.createTypeDeclaration(typeName)
-
-                apiModifier.setTypeDeclarationSchema(bodyWrapper, this.name)
-                apiModifier.setTypeDeclarationExample(bodyWrapper, "!include ./examples/" + this.name + ".json")
-                assistUtils.createGlobalSchemaFromNameAndContent(this.h.root(), this.name, "schemas/" + this.name + ".json", this.jsschema)
-                saveExample(this.h, "./examples/" + this.name + ".json", this.jsexample);
-
-                bodies.push(bodyWrapper);
-            }
-
-            if (this.needXML) {
-                var typeName = "application/xml"
-
-                var bodyWrapper = apiModifier.createTypeDeclaration(typeName)
-
-                apiModifier.setTypeDeclarationSchema(bodyWrapper, this.name + "-xml")
-                apiModifier.setTypeDeclarationExample(bodyWrapper, "!include ./examples/" + this.name + ".xml")
-                var xmlSchemaContents = this.xmlschema;
-                assistUtils.createGlobalSchemaFromNameAndContent(this.h.root(), this.name + "-xml", "schemas/" + this.name + ".xsd", xmlSchemaContents)
-                saveExample(this.h, "./examples/" + this.name + ".xml", this.xmlexample);
-
-                bodies.push(bodyWrapper);
-            }
-
-            bodies.forEach(bodyWrapper => {
-                var foundWrapper = _.find(responseWrapper.body() || [], foundWrapper => bodyWrapper.name() === foundWrapper.name());if(foundWrapper) {
-                    (<any>responseWrapper).remove(foundWrapper);
-                }
-
-                apiModifier.addChild(responseWrapper, bodyWrapper);
-            })
-        }
-
-        var rs = this.h.lowLevel().unit().contents();
-        assistUtils.getActiveEditor().setText(assistUtils.cleanEmptyLines(rs));
-    }
-}
-
-class ExtractResourceTypeStateCalculator extends commonContextActions.CommonASTStateCalculator {
+class ExtractResourceTypeStateCalculator extends contextActions.CommonASTStateCalculator {
 
     calculate():any {
 
@@ -807,7 +502,7 @@ class ExtractResourceTypeStateCalculator extends commonContextActions.CommonASTS
     }
 }
 
-class ExtractTraitStateCalculator extends commonContextActions.CommonASTStateCalculator {
+class ExtractTraitStateCalculator extends contextActions.CommonASTStateCalculator {
 
     calculate():any {
 
@@ -827,7 +522,7 @@ class ExtractTraitStateCalculator extends commonContextActions.CommonASTStateCal
     }
 }
 
-class MoveContentStateCalculator extends commonContextActions.CommonASTStateCalculator {
+class MoveContentStateCalculator extends contextActions.CommonASTStateCalculator {
 
     calculate():any {
 
@@ -847,7 +542,7 @@ class MoveContentStateCalculator extends commonContextActions.CommonASTStateCalc
     }
 }
 
-class ConvertJsonSchemaToTypeStateCalculator extends commonContextActions.CommonASTStateCalculator {
+class ConvertJsonSchemaToTypeStateCalculator extends contextActions.CommonASTStateCalculator {
     calculate():any {
         var generalState = this.getGeneralState()
         if (!generalState) return null
@@ -884,180 +579,40 @@ class ConvertJsonSchemaToTypeStateCalculator extends commonContextActions.Common
 }
 
 
-class ConvertTypeToJsonSchemaStateCalculator extends commonContextActions.CommonASTStateCalculator {
-    calculate():any {
-        var generalState = this.getGeneralState()
-        if (!generalState) return null
-        var highLevelNode = <hl.IHighLevelNode>generalState.node;
-        //console.log('definition: ' + highLevelNode.definition().name() + '; ' + generalState.completionKind);
-        if (generalState.completionKind != search.LocationKind.KEY_COMPLETION
-            && generalState.completionKind != search.LocationKind.VALUE_COMPLETION)
-            return null;
-        var txt = generalState.editor.getText();
-        //var res = getKeyValue(generalState.offset, txt);
-        //return (res == 'type')? highLevelNode: null;
-
-        var attr = _.find(highLevelNode.attrs(),
-            x=>x.lowLevel().start() < generalState.offset && x.lowLevel().end() >= generalState.offset && !x.property().getAdapter(def.RAMLPropertyService).isKey());
-
-        if (!attr) return null
-
-        if (!attr.value()) return null
-
-        var p:hl.IProperty = attr.property();
-
-        if (!universeHelpers.isTypeProperty(p)) return null
-        return highLevelNode
-    }
-}
-
-class ConvertTypeToJsonSchemaAtTypeStateCalculator extends commonContextActions.CommonASTStateCalculator {
-    calculate():any {
-        var generalState = this.getGeneralState()
-        if (!generalState) return null
-        var node = <hl.IHighLevelNode>generalState.node;
-        //highLevelNode.lowLevel().show('HL');
-        //console.log('node def: ' + node.property().name() + ': ' + node.definition().name() + '; ' + generalState.completionKind);
-        if (generalState.completionKind != search.LocationKind.SEQUENCE_KEY_COPLETION)
-            return null;
-        return universeHelpers.isTypesProperty(node.property()) ? node : null;
-    }
-}
-
-function indent(line:string) {
-    var rs = "";
-    for (var i = 0; i < line.length; i++) {
-        var c = line[i];
-        if (c == '\r' || c == '\n') {
-            continue;
-        }
-        if (c == ' ' || c == '\t') {
-            rs += c;
-            continue;
-        }
-        break;
-    }
-    return rs;
-}
+// function indent(line:string) {
+//     var rs = "";
+//     for (var i = 0; i < line.length; i++) {
+//         var c = line[i];
+//         if (c == '\r' || c == '\n') {
+//             continue;
+//         }
+//         if (c == ' ' || c == '\t') {
+//             rs += c;
+//             continue;
+//         }
+//         break;
+//     }
+//     return rs;
+// }
 
 //FIXME remove it from here duplication with jsyaml2lowLevel.ts
-function stripIndent(text:string, indent:string) {
-    var lines = assistUtils.splitOnLines(text);
-    var rs = [];
-    for (var i = 0; i < lines.length; i++) {
-        if (i == 0) {
-            rs.push(lines[0]);
-        }
-        else {
-            rs.push(lines[i].substring(indent.length));
-        }
-    }
-    return rs.join("");
-}
+// function stripIndent(text:string, indent:string) {
+//     var lines = assistUtils.splitOnLines(text);
+//     var rs = [];
+//     for (var i = 0; i < lines.length; i++) {
+//         if (i == 0) {
+//             rs.push(lines[0]);
+//         }
+//         else {
+//             rs.push(lines[i].substring(indent.length));
+//         }
+//     }
+//     return rs.join("");
+// }
 
-export class MoveToNewFileDialog {
 
-    constructor(private node:hl.IHighLevelNode) {
 
-    }
-
-    destination:string;
-
-    show() {
-        var zz:any = null;
-        var node = this.node;
-        var vc = UI.section("Move node content to new file ", UI.Icon.GIST_NEW, false, false);
-        var errorLabel = UI.label("Please enter destination file path", UI.Icon.BUG, UI.TextClasses.ERROR, UI.HighLightClasses.NONE);
-        vc.addChild(UI.vc(errorLabel));
-        vc.addChild(UI.label("Please enter destination path"));
-        var txt = UI.texfField("", "", x=> {
-            if (!txt) {
-                return;
-            }
-
-            var errorMessage = null;
-
-            this.destination = txt.getBinding().get();
-            if (this.destination.trim().length == 0) {
-                errorMessage = "Please enter destination file path";
-            }
-            else if (!path.extname(this.destination) || path.extname(this.destination).trim().length <= 2) {
-                errorMessage = "Please enter destination file extension";
-            }
-            else {
-                var dir = path.resolve(path.dirname(assistUtils.getActiveEditor().getPath()), path.dirname(this.destination));
-                if (!fs.existsSync(dir)) {
-                    errorMessage = "Parent directory does not exist"
-                }
-                else {
-                    var st = fs.statSync(dir)
-                    if (!st.isDirectory()) {
-                        errorMessage = "Parent path is not a directory"
-                    }
-
-                    if(!errorMessage) {
-                        var canWrite = true;
-
-                        try {
-                            (<any>fs).accessSync(dir, (<any>fs).W_OK);
-                        } catch(exception) {
-                            canWrite = false;
-                        }
-
-                        if(!canWrite) {
-                            errorMessage = "Can't write to specified directory, access denied. Please, check your permissions."
-                        }
-                    }
-                }
-            }
-
-            if (errorMessage) {
-                errorLabel.setDisplay(true);
-                errorLabel.setText(errorMessage);
-                okButton.setDisabled(true);
-            } else {
-                errorLabel.setDisplay(false);
-                okButton.setDisabled(false);
-            }
-
-        });
-        vc.addChild(UI.vc(txt));
-        var buttonBar = UI.hc().setPercentWidth(100).setStyle("display", "flex");
-        buttonBar.addChild(UI.label("", null, null, null).setStyle("flex", "1"))
-        buttonBar.addChild(UI.button("Cancel", UI.ButtonSizes.NORMAL, UI.ButtonHighlights.NO_HIGHLIGHT, UI.Icon.NONE, x=> {
-            zz.destroy()
-        }).margin(10, 10))
-        var okButton = UI.button("Move", UI.ButtonSizes.NORMAL, UI.ButtonHighlights.SUCCESS, UI.Icon.NONE, x=> {
-            var d = path.resolve(path.dirname(assistUtils.getActiveEditor().getPath()), this.destination);
-            var dump = this.node.lowLevel().dump();
-            var ci = assistUtils.splitOnLines(dump);
-            var li = ci.length > 1 ? indent(ci[1]) : indent(ci[0]);
-            dump = dump.substring(this.node.lowLevel().keyEnd() - this.node.lowLevel().start() + 1).trim();
-            dump = stripIndent(dump, li);
-
-            var ramlComment = node.definition().universe().version()==="RAML10" ? "#%RAML 1.0 " : "#%RAML 0.8 "
-
-            dump = ramlComment + this.node.definition().nameId() + "\n" + dump;
-            fs.writeFileSync(d, dump);
-            //no we need to replace content of the node with text;
-
-            var txt = node.lowLevel().unit().contents();
-            var endPart = txt.substring(node.lowLevel().end());
-            var startPart = txt.substring(0, node.lowLevel().keyEnd() + 1);
-            var vl = startPart + " !include " + this.destination + endPart;
-            assistUtils.getActiveEditor().setText(vl);
-            zz.destroy();
-        });
-        okButton.setDisabled(true)
-        buttonBar.addChild(okButton);
-        vc.addChild(buttonBar)
-        var html = vc.renderUI();
-        zz = (<any>atom).workspace.addModalPanel({item: html});
-        html.focus();
-    }
-}
-
-class ExtractLibraryStateCalculator extends commonContextActions.CommonASTStateCalculator {
+class ExtractLibraryStateCalculator extends contextActions.CommonASTStateCalculator {
 
     calculate():any {
 
@@ -1072,7 +627,7 @@ class ExtractLibraryStateCalculator extends commonContextActions.CommonASTStateC
     }
 }
 
-class ExtractOverlayStateCalculator extends commonContextActions.CommonASTStateCalculator {
+class ExtractOverlayStateCalculator extends contextActions.CommonASTStateCalculator {
 
     calculate():any {
 
@@ -1087,7 +642,7 @@ class ExtractOverlayStateCalculator extends commonContextActions.CommonASTStateC
     }
 }
 
-class ModifyOverlayStateCalculator extends commonContextActions.CommonASTStateCalculator {
+class ModifyOverlayStateCalculator extends contextActions.CommonASTStateCalculator {
 
     calculate():any {
 
@@ -1108,7 +663,7 @@ class ModifyOverlayStateCalculator extends commonContextActions.CommonASTStateCa
     }
 }
 
-class AbstractMoveTypePropertiesCalculator extends commonContextActions.CommonASTStateCalculator {
+class AbstractMoveTypePropertiesCalculator extends contextActions.CommonASTStateCalculator {
     isTypeNode(node) {
         return (<hl.IHighLevelNode>node).property
             && (<hl.IHighLevelNode>node).property().range
@@ -1133,7 +688,7 @@ class PullUpStateCalculator extends AbstractMoveTypePropertiesCalculator {
                 //so, we're inside the type
 
                 //but we also want to double check that there are superclasses
-                var superTypes = extract.findUserDefinedSupertypes(<hl.IHighLevelNode>current)
+                var superTypes = extract.findUserDefinedSupertypes(hlConv(<hl.IHighLevelNode>current))
                 if (superTypes && superTypes.length > 0)
                     return current
                 else
@@ -1161,7 +716,7 @@ class ExtractSupertypeCalculator extends AbstractMoveTypePropertiesCalculator {
                 //so, we're inside the type
 
                 //but we also want to double check that there are no superclasses
-                var superTypes = extract.findUserDefinedSupertypes(<hl.IHighLevelNode>current)
+                var superTypes = extract.findUserDefinedSupertypes(hlConv(<hl.IHighLevelNode>current))
                 if (!superTypes || superTypes.length == 0)
                     return current
                 else
@@ -1174,20 +729,7 @@ class ExtractSupertypeCalculator extends AbstractMoveTypePropertiesCalculator {
     }
 }
 
-
-class CommentNodeCalculator extends commonContextActions.CommonASTStateCalculator {
-
-    calculate():any {
-
-        var generalState = this.getGeneralState()
-        if (!generalState) return null
-
-        return generalState
-    }
-}
-
-
-class GenerateExampleCalculator extends commonContextActions.CommonASTStateCalculator {
+class GenerateExampleCalculator extends contextActions.CommonASTStateCalculator {
     calculate():any {
         var generalState = this.getGeneralState()
         if (!generalState) return null
@@ -1214,37 +756,6 @@ class GenerateExampleCalculator extends commonContextActions.CommonASTStateCalcu
 
         return highLevelNode
     }
-}
-
-export function findLowLevelNodeByOffset(root:lowLevel.ILowLevelASTNode, offset:number):lowLevel.ILowLevelASTNode {
-    if ((root.keyStart() > offset || root.valueEnd() < offset) && root.parent()) {
-        return null;
-    }
-
-    if(root.includedFrom()) {
-        return findLowLevelNodeByOffset(root.includedFrom(), offset);
-    }
-
-    var children = root.children()
-    for (var key in children) {
-        var child = children[key]
-        var result = findLowLevelNodeByOffset(child, offset)
-        if (result) return result;
-    }
-
-    return root;
-}
-
-function lastChild(root:lowLevel.ILowLevelASTNode) {
-    if(root.includedFrom()) {
-        return root.includedFrom();
-    }
-
-    if(!root.children() || root.children().length === 0) {
-        return root;
-    }
-
-    return lastChild(root.children().filter(child => child ? true : false)[root.children().length - 1]);
 }
 
 export function initialize() {
@@ -1274,40 +785,12 @@ export function initialize() {
 
         onClick: (state)=> {
             new move.MoveElementsDialog(
-                <hl.IHighLevelNode>state,
+                hlConv(<hl.IHighLevelNode>state),
                 "Resource Type", true).show()
         },
 
         stateCalculator: new MoveResourceStateCalculator(),
 
-        shouldDisplay: state=>state != null
-    })
-
-    //contextActions.addAction({
-    //
-    //    name: "Expand Signature",
-    //
-    //    target: contextActions.TARGET_RAML_EDITOR_NODE,
-    //
-    //    category: ["Refactoring"],
-    //
-    //    onClick: state=>expandSignature(<hl.IAttribute> state),
-    //
-    //    stateCalculator: new ExpandSignatureStateCalculator(),
-    //
-    //    shouldDisplay: state=>state != null
-    //})
-
-    contextActions.addAction({
-        name: "Complete body",
-        target: contextActions.TARGET_RAML_EDITOR_NODE,
-        category: ["Add new..."],
-        onClick: (state)=> {
-            var h = <hl.IHighLevelNode>state;
-            h.lowLevel().show('BODY');
-            new FillBodyDialog(h.parent().parent(), h).show()
-        },
-        stateCalculator: new CompleteBodyStateCalculator(),
         shouldDisplay: state=>state != null
     })
 
@@ -1320,7 +803,7 @@ export function initialize() {
         category: ["Refactoring"],
 
         onClick: (state)=> {
-            new extract.ExtractTypesAndTraitsDialog(<hl.IHighLevelNode>state, "Resource Type", true).show()
+            new extract.ExtractTypesAndTraitsDialog(hlConv(<hl.IHighLevelNode>state), "Resource Type", true).show()
         },
 
         stateCalculator: new ExtractResourceTypeStateCalculator(),
@@ -1337,27 +820,10 @@ export function initialize() {
         category: ["Refactoring"],
 
         onClick: (state)=> {
-            new extract.ExtractTypesAndTraitsDialog(<hl.IHighLevelNode>state, "Trait", false).show()
+            new extract.ExtractTypesAndTraitsDialog(hlConv(<hl.IHighLevelNode>state), "Trait", false).show()
         },
 
         stateCalculator: new ExtractTraitStateCalculator(),
-
-        shouldDisplay: state=>state != null
-    });
-
-    contextActions.addAction({
-
-        name: "Move content to other file",
-
-        target: contextActions.TARGET_RAML_EDITOR_NODE,
-
-        category: ["Refactoring"],
-
-        onClick: (state)=> {
-            new MoveToNewFileDialog(<hl.IHighLevelNode>state).show()
-        },
-
-        stateCalculator: new MoveContentStateCalculator(),
 
         shouldDisplay: state=>state != null
     });
@@ -1372,53 +838,7 @@ export function initialize() {
         stateCalculator: new ConvertJsonSchemaToTypeStateCalculator(),
         shouldDisplay: state=>state != null
     });
-
-    contextActions.addAction({
-        name: "Expand type to JSON schema",
-        target: contextActions.TARGET_RAML_EDITOR_NODE,
-        category: ["Refactoring"],
-        onClick: (state)=> {
-            var node = <hl.IHighLevelNode>state;
-            var api = node.root();
-            var type = node.attrValue('type');
-            //console.log('schema: ' + schema);
-            var types = <hl.IHighLevelNode[]>api.elementsOfKind('types');
-            var typeNode = _.find(types, y=>y.name() == type);
-            if (typeNode) {
-                node.attr('type').setValue('');
-                var obj = su.createModelToSchemaGenerator().generateSchema(typeNode);
-                var text = JSON.stringify(obj, null, 2);
-                node.attrOrCreate('schema').setValue(text);
-                text = api.lowLevel().unit().contents();
-                assistUtils.getActiveEditor().setText(text);
-            }
-        },
-        stateCalculator: new ConvertTypeToJsonSchemaStateCalculator(),
-        shouldDisplay: state=>state != null
-    });
-    contextActions.addAction({
-        name: "Expand type to JSON schema definition",
-        target: contextActions.TARGET_RAML_EDITOR_NODE,
-        category: ["Refactoring"],
-        onClick: (state)=> {
-            var typenode = <hl.IHighLevelNode>state;
-            var api = typenode.root();
-            //console.log('generate type ' + typenode.name());
-            var obj = su.createModelToSchemaGenerator().generateSchema(typenode);
-            var schema = JSON.stringify(obj, null, 2);
-            console.log('schema: ' + schema);
-            //schema = textutil.fromMutiLine(schema);
-            var schemaStub = stubs.createStubNoParentPatch(api, 'schemas', typenode.name());
-            schemaStub.attrOrCreate('value').setValue(schema);
-            api.add(schemaStub);
-            var text = api.lowLevel().unit().contents();
-            //console.log('text:\n' + text);
-            assistUtils.getActiveEditor().setText(text);
-        },
-        stateCalculator: new ConvertTypeToJsonSchemaAtTypeStateCalculator(),
-        shouldDisplay: state=>state != null
-    });
-
+    
     contextActions.addAction({
 
         name: "Extract Library",
@@ -1428,7 +848,7 @@ export function initialize() {
         category: ["Refactoring"],
 
         onClick: (state)=> {
-            new extract.ExtractLibraryDialog(<hl.IHighLevelNode>state, "Extract Library").show()
+            new extract.ExtractLibraryDialog(hlConv(<hl.IHighLevelNode>state), "Extract Library").show()
         },
 
         stateCalculator: new ExtractLibraryStateCalculator(),
@@ -1445,7 +865,7 @@ export function initialize() {
         category: ["Code"],
 
         onClick: (state)=> {
-            new extract.ExtractOverlayDialog(<hl.IHighLevelNode>state, "Extract Overlay").show()
+            new extract.ExtractOverlayDialog(hlConv(<hl.IHighLevelNode>state), "Extract Overlay").show()
         },
 
         stateCalculator: new ExtractOverlayStateCalculator(),
@@ -1462,7 +882,7 @@ export function initialize() {
         category: ["Code"],
 
         onClick: (state)=> {
-            new extract.ModifyOverlayDialog(<hl.IHighLevelNode>state, "Modify Overlay").show()
+            new extract.ModifyOverlayDialog(hlConv(<hl.IHighLevelNode>state), "Modify Overlay").show()
         },
 
         stateCalculator: new ModifyOverlayStateCalculator(),
@@ -1479,7 +899,7 @@ export function initialize() {
         category: ["Refactoring"],
 
         onClick: (state)=> {
-            new extract.PullUpDialog(<hl.IHighLevelNode>state, "Pull Up").show()
+            new extract.PullUpDialog(hlConv(<hl.IHighLevelNode>state), "Pull Up").show()
         },
 
         stateCalculator: new PullUpStateCalculator(),
@@ -1496,88 +916,12 @@ export function initialize() {
         category: ["Refactoring"],
 
         onClick: (state)=> {
-            new extract.ExtractSupertypeDialog(<hl.IHighLevelNode>state, "Extract Supertype").show()
+            new extract.ExtractSupertypeDialog(hlConv(<hl.IHighLevelNode>state), "Extract Supertype").show()
         },
 
         stateCalculator: new ExtractSupertypeCalculator(),
 
         shouldDisplay: state=>state != null
     })
-
-    contextActions.addAction({
-
-        name: "Comment node",
-
-        target: contextActions.TARGET_RAML_EDITOR_NODE,
-
-        category: ["Code"],
-
-        onClick: (state)=> {
-            var highLevelNode:hl.IParseResult = (<commonContextActions.IGeneralASTState>state).node;
-
-            if (!highLevelNode.lowLevel()) return;
-
-            var lowLevelNode = findLowLevelNodeByOffset(highLevelNode.lowLevel(),
-                (<commonContextActions.IGeneralASTState>state).offset)
-
-            if (!lowLevelNode) return;
-
-            var startOffset = lowLevelNode.keyStart() > -1 ? lowLevelNode.keyStart() : lowLevelNode.start();
-
-            lowLevelNode = lastChild(lowLevelNode);
-
-            var endOffset = lowLevelNode.valueEnd() > -1 ? lowLevelNode.valueEnd() : lowLevelNode.end();
-
-            var buffer = (<commonContextActions.IGeneralASTState>state)
-                .editor.getBuffer();
-            var startPosition = buffer.positionForCharacterIndex(startOffset);
-            var startLine = startPosition.row;
-
-            var endPosition = buffer.positionForCharacterIndex(endOffset);
-            var endLine = endPosition.row;
-
-            for (var lineNumber:number = startLine; lineNumber <= endLine; lineNumber++) {
-
-                var oldRange = buffer.rangeForRow(lineNumber, true);
-                var oldText = buffer.getTextInRange(oldRange);
-                var newText = "#" + oldText;
-
-                buffer.setTextInRange(oldRange, newText)
-            }
-        },
-
-        stateCalculator: new CommentNodeCalculator(),
-
-        shouldDisplay: state=>state != null
-    })
-
-    // contextActions.addAction({
-    //     name: "Generate example",
-    //     target: contextActions.TARGET_RAML_EDITOR_NODE,
-    //     category: ["Code"],
-    //     onClick: (state)=> {
-    //         var node = <hl.IHighLevelNode>state;
-    //         var api = node.root();
-    //         var typeman = new gu.TypeManager(<wrapper.ApiImpl>api.wrapperNode());
-    //         var type = node.attr('type').value();
-    //         var egen = new genex.ExampleGenerator(typeman);
-    //         var nodetype = node.definition().nameId();
-    //         var proptype = node.property().range().nameId();
-    //         if(nodetype != 'application/json') {
-    //             type = node.name();
-    //         }
-    //         //console.log('node type: ' + nodetype + '; prop type: ' + proptype + ' ==> ' + type);
-    //         var json = egen.generateTypeExpression(type);
-    //         var example = JSON.stringify(json, null, 2);
-    //         node.attrOrCreate('example').setValue(example);
-    //         var text = api.lowLevel().unit().contents();
-    //         //console.log('text:\n' + text);
-    //         assistUtils.getActiveEditor().setText(text);
-    //         //new extract.ExtractSupertypeDialog(<hl.IHighLevelNode>state, "Extract Supertype").show()
-    //     },
-    //     stateCalculator: new GenerateExampleCalculator(),
-    //     shouldDisplay: state=>state != null
-    // });
-
 
 }
