+++ src/atom/editor-tools/details2.ts
@@ -17,7 +17,7 @@ import shemagen=require("../../util/schemaGenerator")
 import SpacePenViews = require('atom-space-pen-views')
 import move=require("../dialogs/moveElementsDialog")
 import tooltip=require("../core/tooltip-manager")
-import contextActions = require("../context-menu/contextActions")
+import contextActions = require("raml-actions")
 import commonContextActions = require("../context-menu/commonContextActions")
 import suggestions=require("./suggestions-panel")
 import assistUtils = require("../dialogs/assist-utils");
