+++ src/atom/editor-tools/schemaUI.ts
@@ -10,7 +10,7 @@ import fs=require("fs")
 import Opt=require("../../Opt")
 import assistUtils = require("../dialogs/assist-utils");
 import details2=require("../editor-tools/details2")
-import contextActions = require("../context-menu/contextActions")
+import contextActions = require("raml-actions")
 import commonContextActions = require("../context-menu/commonContextActions")
 import details=require("../editor-tools/details")
 import _=require("underscore")
